+++ include/fsl_esdhc.h
@@ -155,8 +155,8 @@
 #define ESDHC_HOSTCAPBLT_HSS	0x00200000
 
 struct fsl_esdhc_cfg {
-	u32	esdhc_base;
-	u32	sdhc_clk;
+	uintptr_t esdhc_base;
+	uintptr_t sdhc_clk;
 	u8	max_bus_width;
 	struct mmc_config cfg;
 };
